Don't use raw requires where unnecessary in tailwind.config.cjs
Using raw require results in violating `eslint@typescript-eslint/no-var-requires`, which is one of the rules that comes enabled by default in this template. We can fix this by using import statements where possible. We still need to keep the `require.resolve` however, so this is not a full transition to ESModule syntax.